requirements.txt
tensorflow==2.15.0
flask
numpy
pandas
matplotlib
scikit-learn
Pillow
PyYAML
gdown
